* ext/date/date_strftime.c (date_strftime_wo_timespec):   surpress warning: shorten-64-to-32.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@31144 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/date/date_strftime.c
@@ -491,9 +491,9 @@ date_strftime_wo_timespec(char *s, size_t maxsize, const char *format,
 
 				off = NUM2LONG(rb_funcall(vtm->utc_offset, rb_intern("round"), 0));
 
-				aoff = off;
+				aoff = (int)off;
 				if (aoff < 0)
-					aoff = -off;
+					aoff = (int)-off;
 
 				if ((aoff / 3600) < 10)
 					hl = 1;
